[Chore] #189 - DetailSearchBottomView 생성하여 뷰 로직 분리
