+++ WSSiOS/WSSiOS/Source/Presentation/Search/DetailSearch/DetailSearchView/DetailSearchAssistantView/DetailSearchHeaderView.swift
@@ -1,0 +1,8 @@
+//
+//  DetailSearchHeaderView.swift
+//  WSSiOS
+//
+//  Created by Seoyeon Choi on 7/19/24.
+//
+
+import Foundation
